+++ astroquery/esa/euclid/core.py
@@ -11,12 +11,13 @@ import os
 import pprint
 import tarfile
 import zipfile
+from collections.abc import Iterable
+from datetime import datetime
+
 from astropy import units
 from astropy import units as u
 from astropy.coordinates import Angle
 from astropy.units import Quantity
-from collections.abc import Iterable
-from datetime import datetime
 from requests.exceptions import HTTPError
 
 from astroquery import log
@@ -142,12 +143,14 @@ class EuclidClass(TapPlus):
             the ‘ra’ column in the table table_a_full_qualified_name
         table_a_column_dec :  str, mandatory
             the ‘dec’ column in the table table_a_full_qualified_name
-        table_b_full_qualified_name : str, optional, default the main_table associated to the selected environment
-            a full qualified table name (i.e. schema name and table name)
-        table_b_column_ra : str, optional, default the main_table_ra_column associated to the selected environment
-            the ‘ra’ column in the table table_b_full_qualified_name
-        table_b_column_dec :  str, default the main_table_dec_column associated to the selected environment
-            the ‘dec’ column in the table table_b_full_qualified_name
+        table_b_full_qualified_name : str, optional, default the main_table associated to the selected environment a
+            full qualified table name (i.e. schema name and table name). 'table_name' and 'table_b_column_ra' and
+            'table_b_column_dec' are independent.
+        table_b_column_ra : str, optional, default the main_table_ra_column associated to the selected environment the
+            ‘ra’ column in the table table_b_full_qualified_name. 'table_b_column_ra' and 'table_b_column_dec' are not
+            independent.
+        table_b_column_dec :  str, default the main_table_dec_column associated to the selected environment the ‘dec’
+            column in the table table_b_full_qualified_name
         results_name : str, optional, default None
             custom name defined by the user for the job that is going to be created
         radius : float (arc. seconds), str or astropy.coordinate, optional, default 1.0
@@ -175,13 +178,17 @@ class EuclidClass(TapPlus):
         if not schema_a:
             raise ValueError(f"Schema name is empty in full qualified table: '{table_a_full_qualified_name}'")
 
-        if table_b_full_qualified_name is None:
+        if table_b_full_qualified_name is None or not table_b_full_qualified_name:
             table_b_full_qualified_name = self.main_table
+
+        if (table_b_column_ra is None or not table_b_column_ra) and (
+                table_b_column_dec is None or not table_b_column_dec):
             table_b_column_ra = self.main_table_ra
             table_b_column_dec = self.main_table_dec
-        else:
-            if table_b_column_ra is None or table_b_column_dec is None:
-                raise ValueError(f"Invalid ra or dec column names: '{table_b_column_ra}' and '{table_b_column_dec}'")
+
+        if ((table_b_column_ra is None or not table_b_column_ra) and table_b_column_dec is not None) or (
+                table_b_column_ra is not None and (table_b_column_dec is None or not table_b_column_dec)):
+            raise ValueError(f"Invalid ra or dec column names: '{table_b_column_ra}' and '{table_b_column_dec}'")
 
         schema_b = self.__get_schema_name(table_b_full_qualified_name)
         if not schema_b:
@@ -266,7 +273,8 @@ class EuclidClass(TapPlus):
             raise ValueError(f"Not found schema name in full qualified table: '{full_qualified_table_name}'")
         return schema
 
-    def launch_job(self, query, *, name=None, dump_to_file=False, output_file=None, output_format="csv", verbose=False,
+    def launch_job(self, query, *, name=None, dump_to_file=False, output_file=None, output_format="votable_gzip",
+                   verbose=False,
                    upload_resource=None, upload_table_name=None):
         """
         Launches a synchronous job
@@ -282,8 +290,9 @@ class EuclidClass(TapPlus):
         output_file : str, optional, default None
             File name where the results are saved if dump_to_file is True.
             If this parameter is not provided, the job id is used instead
-        output_format : str, optional, default 'csv'
-            output format for the output file
+        output_format : str, optional, default 'votable_gzip'
+            output format for the output file. Available formats are: 'votable', 'votable_plain', 'fits', 'csv',
+            'ecsv' and 'json', default is a compressed 'votable'
         verbose : bool, optional, default 'False'
             flag to display information about the process
         upload_resource: str, optional, default None
@@ -310,7 +319,7 @@ class EuclidClass(TapPlus):
         except Exception as exx:
             log.error(f'Query failed: {query}, {str(exx)}')
 
-    def launch_job_async(self, query, *, name=None, dump_to_file=False, output_file=None, output_format="csv",
+    def launch_job_async(self, query, *, name=None, dump_to_file=False, output_file=None, output_format="votable_gzip",
                          verbose=False, background=False, upload_resource=None, upload_table_name=None, autorun=True):
         """
         Launches an asynchronous job
@@ -326,8 +335,9 @@ class EuclidClass(TapPlus):
         output_file : str, optional, default None
             file name where the results are saved if dump_to_file is True.
             if this parameter is not provided, the jobid is used instead
-        output_format : str, optional, default 'csv'
-            format of the results for the output file
+        output_format : str, optional, default 'votable_gzip'
+            format of the results for the output file. Available formats are: 'votable', 'votable_plain', fits',
+            'csv', 'ecsv' and 'json', default is a compressed 'votable'
         verbose : bool, optional, default 'False'
             flag to display information about the process
         background : bool, optional, default 'False'
@@ -427,9 +437,10 @@ class EuclidClass(TapPlus):
             coordinates center point
         radius : astropy.units, mandatory
             radius
-        table_name : str, optional, default main table name doing the cone search against
+        table_name : str, optional, default main table name doing the cone search against. `table_name` and
+            `ra_column_name` and `dec_column_name` are independent.
         ra_column_name : str, optional, default ra column in main table
-            ra column doing the cone search against
+            ra column doing the cone search against. `ra_column_name` and `dec_column_name` are not independent.
         dec_column_name : str, optional, default dec column in main table
             dec column doing the cone search against
         async_job : bool, optional, default 'False'
@@ -437,7 +448,6 @@ class EuclidClass(TapPlus):
             synchronous)
         verbose : bool, optional, default 'False'
             flag to display information about the process
-
         columns: list, optional, default None
             if empty, all columns will be selected
         Returns
@@ -447,8 +457,14 @@ class EuclidClass(TapPlus):
 
         if table_name is None:
             table_name = self.main_table
+
+        if ra_column_name is None and dec_column_name is None:
             ra_column_name = self.main_table_ra
             dec_column_name = self.main_table_dec
+
+        if (ra_column_name is not None and dec_column_name is None) or (
+                ra_column_name is None and dec_column_name is not None):
+            raise ValueError(f"Invalid ra or dec column names: ra, {ra_column_name}, dec, {dec_column_name}")
 
         radius_deg = None
         coord = commons.parse_coordinates(coordinate)
@@ -500,7 +516,7 @@ class EuclidClass(TapPlus):
                     background=False,
                     dump_to_file=False,
                     output_file=None,
-                    output_format="csv",
+                    output_format="votable_gzip",
                     verbose=False,
                     columns=None):
         """
@@ -513,9 +529,9 @@ class EuclidClass(TapPlus):
         radius : astropy.units, mandatory
             radius
         table_name : str, optional, default the table defined for the selected environment
-            Table to search
+            Table to search. 'table_name' and 'ra_column_name' and 'dec_column_name`'are independent.
         ra_column_name : str, optional, default the column name defined for the selected environment
-            Name of the RA column in the table
+            Name of the RA column in the table. 'ra_column_name' and 'dec_column_name' are not independent.
         dec_column_name : str, optional, default the column name defined for the selected environment
             Name of the DEC column in the table
         async_job : bool, optional, default 'False'
@@ -529,8 +545,9 @@ class EuclidClass(TapPlus):
         output_file : str, optional, default None
             file name where the results are saved if dump_to_file is True.
             If this parameter is not provided, the job id is used instead
-        output_format : str, optional, default 'csv'
-            Output format for the output file
+        output_format : str, optional, default 'votable_gzip'
+            Output format for the output file. Available formats are: 'votable', 'votable_plain', 'fits', 'csv',
+            'ecsv' and 'json', default is a compressed 'votable'.
         columns: list, optional, default None
             if empty, all columns will be selected
         verbose : bool, optional, default 'False'
@@ -540,6 +557,7 @@ class EuclidClass(TapPlus):
         -------
         A Job object
         """
+
         radius_deg = None
         coord = commons.parse_coordinates(coordinate)
         ra_hours, dec = commons.coord_to_radec(coord)
@@ -547,10 +565,16 @@ class EuclidClass(TapPlus):
         if radius is not None:
             radius_deg = Angle(self.__get_quantity_input(radius, "radius")).to_value(u.deg)
 
-        if table_name is None:
+        if table_name is None or not table_name:
             table_name = self.main_table
+
+        if (ra_column_name is None or not ra_column_name) and (dec_column_name is None or not dec_column_name):
             ra_column_name = self.main_table_ra
             dec_column_name = self.main_table_dec
+
+        if (ra_column_name is not None and (dec_column_name is None or not dec_column_name)) or (
+                (ra_column_name is None or not ra_column_name) and dec_column_name is not None):
+            raise ValueError(f"Invalid ra or dec column names: '{ra_column_name}' and '{dec_column_name}'")
 
         if columns:
             columns = ','.join(map(str, columns))
@@ -604,12 +628,12 @@ class EuclidClass(TapPlus):
 
         Parameters
         ----------
-        user : str, mandatory if 'file' is not provided, default None
+        user : str, mandatory if 'credentials_file' is not provided, default None
             login name
-        password : str, mandatory if 'file' is not provided, default None
+        password : str, mandatory if 'credentials_file' is not provided, default None
             user password
         credentials_file : str, mandatory if no 'user' & 'password' are provided
-            file containing user and password in two lines
+            the file containing user and password in two lines
         verbose : bool, optional, default 'False'
             flag to display information about the process
 
@@ -857,10 +881,10 @@ class EuclidClass(TapPlus):
             self.__eucliddata.load_data(params_dict=params_dict, output_file=output_file_full_path, verbose=verbose)
         except HTTPError as err:
             log.error(f"Cannot retrieve products for observation {id}. HTTP error: {err}")
-            return
+            return None
         except Exception as exx:
             log.error(f'Cannot retrieve products for observation {id}: {str(exx)}')
-            return
+            return None
 
         files = []
         self.__extract_file(output_file_full_path=output_file_full_path, output_dir=output_dir, files=files)
@@ -1220,10 +1244,10 @@ class EuclidClass(TapPlus):
         except HTTPError as err:
             log.error(
                 f"Cannot retrieve products for file_name {file_name} or product_id {product_id}. HTTP error: {err}")
-            return
+            return None
         except Exception as exx:
             log.error(f"Cannot retrieve products for file_name {file_name} or product_id {product_id}: {str(exx)}")
-            return
+            return None
 
         files = []
         self.__extract_file(output_file_full_path=output_file_full_path, output_dir=output_dir, files=files)
@@ -1284,12 +1308,12 @@ class EuclidClass(TapPlus):
             log.error(
                 f"Cannot retrieve the product for file_path {file_path}, obsId {id}, and collection {instrument}. "
                 f"HTTP error: {err}")
-            return
+            return None
         except Exception as exx:
             log.error(
                 f"Cannot retrieve the product for file_path {file_path}, obsId {id}, and collection {instrument}: "
                 f"{str(exx)}")
-            return
+            return None
 
         files = []
         self.__extract_file(output_file_full_path=output_file_full_path, output_dir=output_dir, files=files)
@@ -1372,10 +1396,10 @@ class EuclidClass(TapPlus):
             self.__eucliddata.load_data(params_dict=params_dict, output_file=output_file_full_path, verbose=verbose)
         except HTTPError as err:
             log.error(f'Cannot retrieve spectrum for source_id {source_id}, schema {schema}. HTTP error: {err}')
-            return
+            return None
         except Exception as exx:
             log.error(f'Cannot retrieve spectrum for source_id {source_id}, schema {schema}: {str(exx)}')
-            return
+            return None
 
         files = []
         self.__extract_file(output_file_full_path=output_file_full_path, output_dir=output_dir, files=files)